Fix scripted test and moved to xerial.sbt.pack
