+++ src/main/scala/xerial/sbt/pack/LaunchScript.scala
@@ -1,5 +1,4 @@
-package xerial.sbt
-
+package xerial.sbt.pack
 
 object LaunchScript {
