+++ src/main/scala/xerial/sbt/pack/package.scala
@@ -1,9 +1,10 @@
-package xerial
+package xerial.sbt
 
-import _root_.sbt._
 import java.io._
 
-package object sbt {
+import sbt._
+
+package object pack {
   implicit class ArchiveFile(f: File) {
     def toList: List[String] = Option(f.getParentFile) match {
       case None => f.getName :: Nil
