+++ src/sbt-test/sbt-pack/build-sbt/build.sbt
@@ -1,4 +1,4 @@
-packAutoSettings
+enablePlugins(PackPlugin)
 
 scalaVersion := "2.11.8"
