+++ src/sbt-test/sbt-pack/convention-full/build.sbt
@@ -1,4 +1,4 @@
-packSettings
+enablePlugins(PackPlugin)
 
 scalaVersion := "2.11.8"
