+++ src/sbt-test/sbt-pack/pack-resources/build.sbt
@@ -1,0 +1,13 @@
+lazy val root = Project(
+  id = "pack-resources",
+  base = file("."),
+  settings = Defaults.defaultSettings ++
+    Seq(
+      scalaVersion := "2.11.8",
+      // Copy files from ${root}/web/... to ${root}/target/pack/web-content...
+      packResourceDir += (baseDirectory.value / "web" -> "web-content"),
+      // custom settings here
+      crossPaths := false
+    )
+)
+.enablePlugins(PackPlugin)
