--- src/sbt-test/sbt-pack/pack-resources/project/Build.scala
@@ -1,19 +1,0 @@
-import sbt._
-import sbt.Keys._
-import xerial.sbt.PackPlugin._
-
-object Build extends sbt.Build {
-
-  lazy val root = Project(
-    id = "pack-resources",
-    base = file("."),
-    settings = Defaults.defaultSettings ++ packAutoSettings ++
-      Seq(
-        scalaVersion := "2.11.8",
-        // Copy files from ${root}/web/... to ${root}/target/pack/web-content...
-        packResourceDir += (baseDirectory.value / "web" -> "web-content"),
-        // custom settings here
-        crossPaths := false
-      )
-  )
-}
