+++ src/sbt-test/sbt-pack/space in path/build.sbt
@@ -1,7 +1,5 @@
-packAutoSettings
+enablePlugins(PackPlugin)
 
 scalaVersion := "2.11.6"
-
-
